Update about.md
I did some heavy edits: grammar, typos, etc. 
I changed the numbering of engagements because I felt it was confusing.  For the audience, I think it's more understandable to list the engagements in chronological order. There's actually no requirement that each resource has to be mapped to an engagement. Our scope around that is flexible.